rm bigip fxiture
Issues:
Fixes #62

Problem: We have many sources of truth repeating themselves
for single components of our system.

Analysis:  Part of the fix is to roll all code specific
test enhancements, like fixtures that are used by pytest
to expose elements of the system under test, like "bigip"
into the codebase that they are exposing (f5-common-python in
this case).