Set sdk deps to pinned version based on agent requirements
Issues:
Fixes #72

Problem:
The agent requires a certain version of the sdk when it is installed.
This same version should be used when running tests. We are setting the
required sdk version to 2.3.1 in accordance with the latest release of
the sdk and the agent.

Analysis:
Modified the setup.py to set f5-sdk to version 2.3.1

Tests:
Installed the f5-openstack-test repo and verified 2.3.1 is installed
version of the sdk.