Newton dependencies need revision
Issues:
Fixes #76

Problem:
The newton branch now exists, but we need to revisit the setup.py
install_requires versions of the dependencies used for testing. This
will be largely based on the upper constraints file published by
openstack.

Analysis:
Updated the versions with respect to the upper constraints file
illustrated in this issue on github.

Tests:
Running tests on my local buildbot sandbox