f5-sdk version needs to be bumped in this repo
Issues:
Fixes #79

Problem:
A new sdk was released and the liberty/mitaka versions of the driver
will be using this new sdk. We need to make the change here and merge it
forward.

Analysis:
Bumped the sdk version to 2.3.2

Tests:
Running tests in the agent and driver repos before merging those PRs to
make the latest beta releases.